added delete requests, missing put requests

+++ music_app/app_views/artists.py
@@ -7,45 +7,61 @@ from .. import models
 
 @csrf_exempt
 @api_view(["GET", "POST"])
+# artists
 def artists(request):
     if request.method not in ('GET', 'POST'):
         return HttpResponse(status=405)
 
-    if request.method == 'GET':
+    if request.method == 'GET':  # GET all artists
         data = list(models.Artist.objects.values())
-        print(data)
-        return JsonResponse(data, safe=False)
+        return JsonResponse(data, safe=False, status=200)
 
-    elif request.method == 'POST':
+    elif request.method == 'POST':  # POST a new artist
+        valid_inputs = []
         artist_data = request.data
+        for key in artist_data.keys():
+            valid_inputs.append(key)
+
+        if valid_inputs != ['name', 'age']:
+            return JsonResponse({"mesagge": "Invalid input"}, safe=False, status=400)
+
         encoded = b64encode(artist_data['name'].encode()).decode('utf-8')
-        # print(request.data['name'])
-        new_artist = models.Artist.objects.create(name=artist_data['name'],
-                                                  identificador=encoded[0:22],
-                                                  age=artist_data['age']
-                                                  #   albums=artists_data['albums'],
-                                                  #   tracks=artists_data['tracks'],
-                                                  #   myself=artists_data['myself'],
-                                                  )
-        new_artist.save()
-        new_artist = models.Artist.objects.filter(id=new_artist.id)
-        data_new_artist = list(new_artist.values())
-        return JsonResponse(data_new_artist, safe=False)
+        exists_artist = models.Artist.objects.filter(identificador=encoded)
+        data_artist = list(exists_artist.values())
+
+        if not data_artist:
+            new_artist = models.Artist.objects.create(name=artist_data['name'],
+                                                      identificador=encoded[0:22],
+                                                      age=artist_data['age']
+                                                      #   albums=artists_data['albums'],
+                                                      #   tracks=artists_data['tracks'],
+                                                      #   myself=artists_data['myself'],
+                                                      )
+            new_artist.save()
+            new_artist = models.Artist.objects.filter(id=new_artist.id)
+            data_new_artist = list(new_artist.values())
+            return JsonResponse(data_new_artist, safe=False, status=201)
+
+        else:
+            return JsonResponse(data_artist, safe=False, status=409)
 
 
-# @csrf_exempt
 @api_view(["GET", "DELETE"])
+# artists/<str:artist_id>
 def artists_detail(request, artist_id):
 
-    # if request.method not in ('GET', 'POST'):
-    #     return HttpResponse(status=405)
+    artist = models.Artist.objects.filter(identificador=artist_id)
+    data_artist = list(artist.values())
 
-    if request.method == 'GET':
-        artist = models.Artist.objects.filter(identificador=artist_id)
-        data_artist = list(artist.values())
-        return JsonResponse(data_artist, safe=False)
+    if request.method == 'GET':  # GET artist with artist_id
+        if data_artist:
+            return JsonResponse(data_artist, safe=False, status=200)
+        else:
+            return JsonResponse({"mesagge": "Artist not found"}, status=404)
 
-    elif request.method == 'DELETE':
-        artist = models.Artist.objects.filter(identificador=artist_id)
-        artist.delete()
-        return JsonResponse({"mesagge": "Artist was deleted"})
+    elif request.method == 'DELETE':  # DELETE artist with artist_id
+        if data_artist:
+            artist.delete()
+            return JsonResponse({"mesagge": "Artist was deleted"}, status=204)
+        else:
+            return JsonResponse({"mesagge": "Artist not found"}, status=404)
